Create Options class
Webpay and Patpass Comercio have different headers name on the API for
same thinks. We dont know if Transbank will fix it some day. By now we
should manage this with an abstract Options class in order to each
product could retrieve right header name when we make the http request.