Add sensible default positional arguments
The InscriptionFinishResponse class expects three positional arguments
without defaults: response_code, tbk_user and authorization_code.
However, in an integration environment, whenever the inscription is not
successful, transbank responds only with the response_code, omitting
tbk_user and authorization code.

For this reason, when instantiating the class inside the "finish" method
of the MallInscription class, the runtime throws a TypeError exception
because it is missing two positional arguments.

This error can be easily managed by catching the exception,
but the response_code will not reach the user code.