Fix duplication on Mall transactions
Every time a user refreshed the view previous to the payment (or
similar), the base amount of the transaction was doubled.

This was fixed by changing the __init__ method of the
MallTransactionCreateDetails class.

The class attribute "__details" was left there in case a user uses the
"add" method without instantiating the class.